+++ libs/vector-map/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Version history for `vector-map`
 
-## 1.1.0.2
+## 1.2.0.0
 
-*
+* Add `mapMaybeWithKey`
+* Add `unionWithKey`, `unionWith` and `union`
+* Switch `Semigroup` instance for `VMap` to be left-biased, just like it is for `Map`
 
 ## 1.1.0.1